+++ src/main/java/dev/jbang/dependencies/MavenRepo.java
@@ -1,5 +1,7 @@
 package dev.jbang.dependencies;
 
+import java.util.Objects;
+
 public class MavenRepo {
 
 	private String id;
@@ -24,6 +26,21 @@ public class MavenRepo {
 
 	public void setUrl(String url) {
 		this.url = url;
+	}
+
+	@Override
+	public boolean equals(Object o) {
+		if (this == o)
+			return true;
+		if (o == null || getClass() != o.getClass())
+			return false;
+		MavenRepo mavenRepo = (MavenRepo) o;
+		return Objects.equals(id, mavenRepo.id) && Objects.equals(url, mavenRepo.url);
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(id, url);
 	}
 
 	@Override
