+++ src/test/java/dev/jbang/source/TestSource.java
@@ -3,7 +3,6 @@ package dev.jbang.source;
 import static dev.jbang.util.Util.writeString;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsInAnyOrder;
-import static org.hamcrest.Matchers.hasItem;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -15,7 +14,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.*;
-import java.util.stream.Collectors;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
@@ -156,7 +154,7 @@ public class TestSource extends BaseTest {
 		Source source = new JavaSource(exampleCommandsWithComments, null);
 		Project prj = source.createProject();
 
-		assertEquals(source.getJavaVersion(), "14+");
+		assertEquals(prj.getJavaVersion(), "14+");
 
 		List<String> deps = prj.getMainSourceSet().getDependencies();
 
@@ -281,45 +279,6 @@ public class TestSource extends BaseTest {
 	}
 
 	@Test
-	void testExtractDependencies() {
-		List<String> deps = Source.extractDependencies("//DEPS blah, blue").collect(Collectors.toList());
-
-		assertTrue(deps.contains("blah"));
-
-		assertTrue(deps.contains("blue"));
-
-	}
-
-	@Test
-	void textExtractRepositories() {
-		List<String> repos = Source	.extractRepositories("//REPOS jcenter=https://xyz.org")
-									.collect(Collectors.toList());
-
-		assertThat(repos, hasItem("jcenter=https://xyz.org"));
-
-		repos = Source	.extractRepositories("//REPOS jcenter=https://xyz.org localMaven xyz=file://~test")
-						.collect(Collectors.toList());
-
-		assertThat(repos, hasItem("jcenter=https://xyz.org"));
-		assertThat(repos, hasItem("localMaven"));
-		assertThat(repos, hasItem("xyz=file://~test"));
-	}
-
-	@Test
-	void textExtractRepositoriesGrape() {
-		List<String> deps = Source.extractRepositories(
-				"@GrabResolver(name=\"restlet.org\", root=\"http://maven.restlet.org\")").collect(Collectors.toList());
-
-		assertThat(deps, hasItem("restlet.org=http://maven.restlet.org"));
-
-		deps = Source	.extractRepositories("@GrabResolver(\"http://maven.restlet.org\")")
-						.collect(Collectors.toList());
-
-		assertThat(deps, hasItem("http://maven.restlet.org"));
-
-	}
-
-	@Test
 	void testExtractOptions() {
 		Source s = new JavaSource(example, null);
 
@@ -340,7 +299,7 @@ public class TestSource extends BaseTest {
 	@Test
 	void testGav() {
 		Source src = new JavaSource(example, null);
-		String gav = src.getGav().get();
+		String gav = src.createProject().getGav().get();
 		assertEquals("org.example:classpath", gav);
 	}
